Modified some uberlazy examples.
In the getfield class of the uberlazy bytecode, calling the function in Helper class to generate a new heap node. Consolidating and resuing code between uberlazy and lazy initialization.

Added an example to test the dynamic dispatch.

Added a function to get the super-most parent class when a list of equivalence objects.